+++ glean-core/rlb/src/test.rs
@@ -20,7 +20,54 @@ const GLOBAL_APPLICATION_ID: &str = "org.mozilla.rlb.test";
 
 // Create a new instance of Glean with a temporary directory.
 // We need to keep the `TempDir` alive, so that it's not deleted before we stop using it.
-fn new_glean() -> tempfile::TempDir {
+fn new_glean(configuration: Option<Configuration>) -> tempfile::TempDir {
+    let dir = tempfile::tempdir().unwrap();
+    let tmpname = dir.path().display().to_string();
+
+    let cfg = match configuration {
+        Some(c) => c,
+        None => Configuration {
+            data_path: tmpname,
+            application_id: GLOBAL_APPLICATION_ID.into(),
+            upload_enabled: true,
+            max_events: None,
+            delay_ping_lifetime_io: false,
+            channel: Some("testing".into()),
+            server_endpoint: Some("invalid-test-host".into()),
+            uploader: None,
+        },
+    };
+
+    crate::reset_glean(cfg, ClientInfoMetrics::unknown(), true);
+    dir
+}
+
+#[test]
+fn send_a_ping() {
+    let _lock = GLOBAL_LOCK.lock().unwrap();
+    env_logger::try_init().ok();
+
+    let (s, r) = crossbeam_channel::bounded::<String>(1);
+
+    // Define a fake uploader that reports back the submission URL
+    // using a crossbeam channel.
+    #[derive(Debug)]
+    pub struct FakeUploader {
+        sender: crossbeam_channel::Sender<String>,
+    };
+    impl net::PingUploader for FakeUploader {
+        fn upload(
+            &self,
+            url: String,
+            _body: Vec<u8>,
+            _headers: Vec<(String, String)>,
+        ) -> net::UploadResult {
+            self.sender.send(url).unwrap();
+            net::UploadResult::HttpStatus(200)
+        }
+    }
+
+    // Create a custom configuration to use a fake uploader.
     let dir = tempfile::tempdir().unwrap();
     let tmpname = dir.path().display().to_string();
 
@@ -31,10 +78,21 @@ fn new_glean() -> tempfile::TempDir {
         max_events: None,
         delay_ping_lifetime_io: false,
         channel: Some("testing".into()),
+        server_endpoint: Some("invalid-test-host".into()),
+        uploader: Some(Box::new(FakeUploader { sender: s })),
     };
 
-    initialize(cfg, ClientInfoMetrics::unknown());
-    dir
+    let _t = new_glean(Some(cfg));
+    crate::dispatcher::block_on_queue();
+
+    // Define a new ping and submit it.
+    const PING_NAME: &str = "test-ping";
+    let custom_ping = private::PingType::new(PING_NAME, true, true, vec![]);
+    custom_ping.submit(None);
+
+    // Wait for the ping to arrive.
+    let url = r.recv().unwrap();
+    assert_eq!(url.contains(PING_NAME), true);
 }
 
 #[test]
@@ -42,7 +100,7 @@ fn disabling_upload_disables_metrics_recording() {
     let _lock = GLOBAL_LOCK.lock().unwrap();
     env_logger::try_init().ok();
 
-    let _t = new_glean();
+    let _t = new_glean(None);
     crate::dispatcher::block_on_queue();
 
     let metric = BooleanMetric::new(CommonMetricData {
@@ -102,12 +160,19 @@ fn initialize_must_not_crash_if_data_dir_is_messed_up() {
         max_events: None,
         delay_ping_lifetime_io: false,
         channel: Some("testing".into()),
+        server_endpoint: Some("invalid-test-host".into()),
+        uploader: None,
     };
 
-    initialize(cfg, ClientInfoMetrics::unknown());
-
-    // TODO: is this test working? is it waiting for init to finish?
-    dispatcher::block_on_queue();
+    reset_glean(cfg, ClientInfoMetrics::unknown(), false);
+    // TODO(bug 1675215): ensure initialize runs through dispatcher.
+    // Glean init is async and, for this test, it bails out early due to
+    // an caused by not being able to create the data dir: we can do nothing
+    // but wait. Tests in other bindings use the dispatcher's test mode, which
+    // runs tasks sequentially on the main thread, so no sleep is required,
+    // because we're guaranteed that, once we reach this point, the full
+    // init potentially ran.
+    std::thread::sleep(std::time::Duration::from_secs(3));
 }
 
 #[test]
@@ -118,20 +183,50 @@ fn queued_recorded_metrics_correctly_record_during_init() {
 
 #[test]
 fn initializing_twice_is_a_noop() {
+    let _lock = GLOBAL_LOCK.lock().unwrap();
+    env_logger::try_init().ok();
+
     let dir = tempfile::tempdir().unwrap();
     let tmpname = dir.path().display().to_string();
 
-    let cfg = Configuration {
-        data_path: tmpname,
-        application_id: GLOBAL_APPLICATION_ID.into(),
-        upload_enabled: true,
-        max_events: None,
-        delay_ping_lifetime_io: false,
-        channel: Some("testing".into()),
-    };
+    reset_glean(
+        Configuration {
+            data_path: tmpname.clone(),
+            application_id: GLOBAL_APPLICATION_ID.into(),
+            upload_enabled: true,
+            max_events: None,
+            delay_ping_lifetime_io: false,
+            channel: Some("testing".into()),
+            server_endpoint: Some("invalid-test-host".into()),
+            uploader: None,
+        },
+        ClientInfoMetrics::unknown(),
+        true,
+    );
 
-    initialize(cfg.clone(), ClientInfoMetrics::unknown());
-    initialize(cfg, ClientInfoMetrics::unknown());
+    dispatcher::block_on_queue();
+
+    reset_glean(
+        Configuration {
+            data_path: tmpname,
+            application_id: GLOBAL_APPLICATION_ID.into(),
+            upload_enabled: true,
+            max_events: None,
+            delay_ping_lifetime_io: false,
+            channel: Some("testing".into()),
+            server_endpoint: Some("invalid-test-host".into()),
+            uploader: None,
+        },
+        ClientInfoMetrics::unknown(),
+        false,
+    );
+
+    // TODO(bug 1675215): ensure initialize runs through dispatcher.
+    // Glean init is async and, for this test, it bails out early due to
+    // being initialized: we can do nothing but wait. Tests in other bindings use
+    // the dispatcher's test mode, which runs tasks sequentially on the main
+    // thread, so no sleep is required. Bug 1675215 might fix this, as well.
+    std::thread::sleep(std::time::Duration::from_secs(3));
 }
 
 #[test]